+++ src/ICESat-2/ATL08.jl
@@ -1,14 +1,40 @@
-function points(granule::ICESat2_Granule{:ATL08}; tracks=icesat2_tracks, step=1, canopy=false, ground=true)
+"""
+    points(g::ICESat2_Granule{:ATL08}; tracks=icesat2_tracks, step=1, canopy=false, ground=true))
+
+Retrieve the points for a given ICESat-2 ATL08 (Land and Vegetation Height) granule as a list of namedtuples, one for each beam.
+The names of the tuples are based on the following fields:
+
+| Column             | Field                                    | Description                                           | Units                        |
+|:------------------ |:---------------------------------------- |:----------------------------------------------------- |:---------------------------- |
+| `longitude`        | `land_segments/longitude`                | Longitude of segment center, WGS84, East=+            | decimal degrees              |
+| `latitude`         | `land_segments/latitude`                 | Latitude of segment center, WGS84, North=+            | decimal degrees              |
+| `height`           | `land_segments/terrain/h_te_mean`        | Standard land-ice segment height                      | m above the WGS 84 ellipsoid |
+| `height_error`     | `land_segments/terrain/h_te_uncertainty` | Total vertical geolocation error                      | m                            |
+| `datetime`         | `land_segments/delta_time`               | + `ancillary_data/atlas_sdp_gps_epoch` + `gps_offset` | date-time                    |
+| `quality`          | `land_segments/terrain_flg`              | Boolean flag indicating the best-quality subset       | 1 = high quality             |
+| `phr`              | `land_segments/ph_removal_flag`          | More than 50% of photons removed                      | -                            |
+| `sensitivity`      | `land_segments/snr`                      | The signal to noise ratio                             | -                            |
+| `scattered`        | `land_segments/msw_flag`                 | Multiple Scattering warning flag                      | -1=unknown; 0=none           |
+| `saturated`        | `land_segments/sat_flag`                 | Saturation detected                                   | -                            |
+| `clouds`           | `land_segments/layer_flag`               | Clouds or blowing snow are likely present             | -                            |
+| `track`            | `gt1l` - `gt3r` groups                   | -                                                     | -                            |
+| `strong_beam`      | `-`                                      | "strong" (true) or "weak" (false) laser power         | -                            |
+| `classification`   | `-`                                      | "ground", "high_canopy"                               | -                            |
+| `height_reference` | `land_segments/dem_h`                    | Height of the (best available) DEM                    | m above the WGS 84 ellipsoid |
+| `detector_id`      | `atlas_spot_number attribute`            | -                                                     | -                            |
+
+You can combine the output in a `DataFrame` with `reduce(vcat, DataFrame.(points(g)))` if you
+want to change the default arguments or `DataFrame(g)` with the default options.
+"""
+function points(granule::ICESat2_Granule{:ATL08}; tracks = icesat2_tracks, step = 1, canopy = false, ground = true)
     dfs = Vector{NamedTuple}()
     HDF5.h5open(granule.url, "r") do file
         t_offset = read(file, "ancillary_data/atlas_sdp_gps_epoch")[1]::Float64 + gps_offset
-        orientation = read(file, "orbit_info/sc_orient")[1]::Int8
 
-        for (i, track) ∈ enumerate(tracks)
-            power = track_power(orientation, track)
+        for track in tracks
             if in(track, keys(file)) && in("land_segments", keys(file[track]))
-                for track_nt ∈ points(granule, file, track, power, t_offset, step, canopy, ground)
-                    track_nt.z[track_nt.z .== fill_value] .= NaN
+                for track_nt in points(granule, file, track, t_offset, step, canopy, ground)
+                    track_nt.height[track_nt.height.==fill_value] .= NaN
                     push!(dfs, track_nt)
                 end
             end
@@ -17,9 +43,17 @@ function points(granule::ICESat2_Granule{:ATL08}; tracks=icesat2_tracks, step=1,
     dfs
 end
 
-function points(::ICESat2_Granule{:ATL08}, file::HDF5.H5DataStore, track::AbstractString, power::AbstractString, t_offset::Float64, step=1, canopy=false, ground=true)
+function points(
+    ::ICESat2_Granule{:ATL08},
+    file::HDF5.H5DataStore,
+    track::AbstractString,
+    t_offset::Float64,
+    step = 1,
+    canopy = false,
+    ground = true,
+)
     if ground
-        zt = file["$track/land_segments/terrain/h_te_median"][1:step:end]::Vector{Float32}
+        zt = file["$track/land_segments/terrain/h_te_mean"][1:step:end]::Vector{Float32}
         tu = file["$track/land_segments/terrain/h_te_uncertainty"][1:step:end]::Vector{Float32}
     end
     if canopy
@@ -37,47 +71,49 @@ function points(::ICESat2_Granule{:ATL08}, file::HDF5.H5DataStore, track::Abstra
     phr = file["$track/land_segments/ph_removal_flag"][1:step:end]::Vector{Int8}
     dem = file["$track/land_segments/dem_h"][1:step:end]::Vector{Float32}
     times = unix2datetime.(t .+ t_offset)
+    atlas_beam_type = attrs(file["$track"])["atlas_beam_type"]::String
+    spot_number = attrs(file["$track"])["atlas_spot_number"]::String
 
     if ground
         gt = (
-            x = x,
-            y = y,
-            z = zt,
-            u = tu,
-            t = times,
-            q = q,
-            phr = Int16.(phr),
+            longitude = x,
+            latitude = y,
+            height = zt,
+            height_error = tu,
+            datetime = times,
+            quality = .!Bool.(q),
+            phr = Bool.(phr),
             sensitivity = sensitivity,
             scattered = Int16.(scattered),
             saturated = Int16.(saturated),
             clouds = Bool.(clouds),
             track = Fill(track, length(times)),
-            power = Fill(power, length(times)),
+            strong_beam = Fill(atlas_beam_type == "strong", length(times)),
             classification = Fill("ground", length(times)),
-            return_number = Fill(2, length(times)),
-            number_of_returns = Fill(2, length(times)),
-            reference = dem,
-            )
+            height_reference = dem,
+            detector_id = Fill(parse(Int8, spot_number), length(times)),
+        )
     end
     if canopy
         ct = (
-            x = x,
-            y = y,
-            z = zc,
-            u = cu,
-            t = times,
-            q = q,
-            phr = Int16.(phr),
+            longitude = x,
+            latitude = y,
+            height = zc,
+            height_error = cu,
+            datetime = times,
+            quality = .!Bool.(q),
+            phr = Bool.(phr),
             sensitivity = sensitivity,
             scattered = Int16.(scattered),
             saturated = Int16.(saturated),
             clouds = Bool.(clouds),
             track = Fill(track, length(times)),
-            power = Fill(power, length(times)),
+            strong_beam = Fill(atlas_beam_type == "strong", length(times)),
             classification = Fill("high_canopy", length(times)),
             return_number = Fill(1, length(times)),
             number_of_returns = Fill(2, length(times)),
-            reference = dem,
+            height_reference = dem,
+            detector_id = Fill(parse(Int8, spot_number), length(times)),
         )
     end
     if canopy && ground
@@ -91,24 +127,24 @@ function points(::ICESat2_Granule{:ATL08}, file::HDF5.H5DataStore, track::Abstra
     end
 end
 
-function lines(granule::ICESat2_Granule{:ATL08}; tracks=icesat2_tracks, step=100, quality=1)
+function lines(granule::ICESat2_Granule{:ATL08}; tracks = icesat2_tracks, step = 100, quality = 1)
     dfs = Vector{NamedTuple}()
     HDF5.h5open(granule.url, "r") do file
         # t_offset = read(file, "ancillary_data/atlas_sdp_gps_epoch")[1]::Float64 + gps_offset
-        orientation = read(file, "orbit_info/sc_orient")[1]::Int8
 
         for track ∈ tracks
-            power = track_power(orientation, track)
             if in(track, keys(file)) && in("land_segments", keys(file[track]))
-                z = file["$track/land_segments/terrain/h_te_median"][1:step:end]::Array{Float32,1}
-                x = file["$track/land_segments/longitude"][1:step:end]::Array{Float32,1}
-                y = file["$track/land_segments/latitude"][1:step:end]::Array{Float32,1}
+                height = file["$track/land_segments/terrain/h_te_mean"][1:step:end]::Array{Float32,1}
+                longitude = file["$track/land_segments/longitude"][1:step:end]::Array{Float32,1}
+                latitude = file["$track/land_segments/latitude"][1:step:end]::Array{Float32,1}
                 # t = file["$track/land_segments/delta_time"][1:step:end]::Array{Float64,1}
                 # times = unix2datetime.(t .+ t_offset)
-                z[z .== fill_value] .= NaN
-                line = makeline(x, y, z)
+                atlas_beam_type = attrs(file["$track"])["atlas_beam_type"]::String
+
+                height[height.==fill_value] .= NaN
+                line = Line(longitude, latitude, height)
                 # i = div(length(t), 2) + 1
-                nt = (geom = line, track = track, power = power, granule = granule.id)
+                nt = (geom = line, track = track, strong_beam = atlas_beam_type == "strong", granule = granule.id)
                 push!(dfs, nt)
             end
         end