+++ src/granule.jl
@@ -1,15 +1,16 @@
 using HDF5
 import Downloads
 
-"""This is a method because it will segfault if precompiled."""
+# This is a method because it will segfault if precompiled.
 function _download(kwargs...)
     downloader = Downloads.Downloader()
-    easy_hook = (easy, _) -> begin
-        Downloads.Curl.setopt(easy, Downloads.Curl.CURLOPT_NETRC, Downloads.Curl.CURL_NETRC_OPTIONAL)
-        Downloads.Curl.setopt(easy, Downloads.Curl.CURLOPT_COOKIEFILE, "")
-    end
+    easy_hook =
+        (easy, _) -> begin
+            Downloads.Curl.setopt(easy, Downloads.Curl.CURLOPT_NETRC, Downloads.Curl.CURL_NETRC_OPTIONAL)
+            Downloads.Curl.setopt(easy, Downloads.Curl.CURLOPT_COOKIEFILE, "")
+        end
     downloader.easy_hook = easy_hook
-    Downloads.download(kwargs...;downloader=downloader)
+    Downloads.download(kwargs...; downloader = downloader)
 end
 
 abstract type Granule end
@@ -18,16 +19,22 @@ function HDF5.h5open(granule::Granule)
     HDF5.h5open(granule.url, "r")
 end
 
-function download!(granule::Granule, folder=".")
+"""
+    download!(granule::Granule, folder=".")
+
+Download the file associated with `granule` to the `folder`, from an http(s) location
+if it doesn't already exists locally.
+
+Will require credentials (netrc) which can be set with [`netrc!`](@ref).
+"""
+function download!(granule::Granule, folder = ".")
     fn = joinpath(abspath(folder), granule.id)
     if isfile(fn)
         granule.url = fn
         return fn
     end
     isfile(granule.url) && return granule
-        if startswith(granule.url, "s3://")
-        download_s3(granule.url, fn)
-    elseif startswith(granule.url, "http")
+    if startswith(granule.url, "http")
         _download(granule.url, fn)
     else
         error("Can't determine how to download $(granule.url)")
@@ -36,6 +43,11 @@ function download!(granule::Granule, folder=".")
     granule
 end
 
+"""
+    rm(granule::Granule)
+
+Remove the file associated with `granule` from the local filesystem.
+"""
 function rm(granule::Granule)
     if isfile(granule.url)
         Base.rm(granule.url)
@@ -43,13 +55,18 @@ function rm(granule::Granule)
         @warn("Can't delete $(granule.url)..")
     end
 end
-    
-function download!(granules::Vector{Granule}, folder::AbstractString)
+
+"""
+    download!(granules::Vector{<:Granule}, folder=".")
+
+Like [`download!`](@ref), but for a vector of `granules`.
+"""
+function download!(granules::Vector{Granule}, folder::AbstractString = ".")
     for granule in granules
         download!(granule, folder)
     end
 end
 
-function filesize(granule::T) where T <: Granule
+function filesize(granule::T) where {T<:Granule}
     filesize(granule.url)
 end