+++ docs/src/changelog.md
@@ -1,7 +1,24 @@
 ## Unreleased
-- Documentation improvements
+- Extents support
+- Search using polygons
+- ICESat-2 S3 access
+- 20m resolution data in ICESat-2 ATL08 v05
+
+## v0.2.0
+
+!!! danger
+    This is a **breaking** release
+
+- Many of the column names have changed to be more descriptive.
+- Documentation and docstring improvements.
+- Tables support, you can now do `DataFrame(granule)`, without having to call `points(granule)`.
+- Memory use improvements, by using SentinelArray of FillArray under the hood.
+- Dropped S3, GeoArrays and LAS/LAZ support.
+- Added GeoInterface support for lines/points and dropped GeoDataFrames
+- Expanded test coverage.
+
 ## v0.1.6
-- Support for ICESat GLAH06 by [alex-s-gardner](https://github.com/alex-s-gardner)
+- Support for ICESat GLAH06 by [alex-s-gardner](https://github.com/alex-s-gardner) :+1::tada:
 
 ## v0.1.5
 - Support for ICESat-2 ATL06